services/fastapi-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.12

# AI & ML
google-generativeai==0.8.3
groq==0.11.0
langchain==0.3.7
langchain-groq==0.2.1
langchain-google-genai==2.0.4
langchain-community==0.3.7
langgraph==0.2.48

# Vector DB
qdrant-client==1.12.1
sentence-transformers==3.3.1

# Databases
motor==3.6.0
redis==5.2.1

# Utilities
httpx==0.27.2
Pillow==11.0.0
pydantic==2.10.0
pydantic-settings==2.6.1

# Pricing KB
pyyaml>=6.0